restored c-bindings crate
