+++ crates/c-bindings/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "c-bindings"
+version = "0.1.0"
+edition = "2021"
+authors = ["Newton Toto <nyachiengatoto@gmail.com>"]
+license = "MIT"
+description = "C Bindings for vach, a pure Rust archive format"
+include = ["src/*", "Cargo.toml"]
+repository = "https://github.com/zeskeertwee/vach"
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+vach = { path = "../vach", features = ["archive", "crypto", "compression"] }
